refactor: split ui into smaller files
