--- Midibard/UI/EnsemblePanel.cs
@@ -1,496 +1,0 @@
-// Copyright (C) 2022 akira0245
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see https://github.com/akira0245/MidiBard/blob/master/LICENSE.
-//
-// This code is written by akira0245 and was originally used in the MidiBard project. Any usage of this code must prominently credit the author, akira0245, and indicate that it was originally used in the MidiBard project.
-
-using System;
-using System.Linq;
-using System.Numerics;
-
-using Dalamud.Interface;
-using Dalamud.Interface.Utility;
-
-using ImGuiNET;
-
-using MidiBard.Control.MidiControl.PlaybackInstance;
-using MidiBard.IPC;
-using MidiBard.Managers;
-using MidiBard.Managers.Ipc;
-using MidiBard.Util;
-
-using static MidiBard2.Resources.Language;
-
-namespace MidiBard;
-
-public partial class PluginUI
-{
-    private bool ShowEnsembleControlWindow;
-
-    private void DrawEnsembleMenu()
-    {
-        var ensembleRunning = MidiBard.AgentMetronome.EnsembleModeRunning;
-        var isEnsembleButtonsDisabled = MidiBard.CurrentPlayback == null || ensembleRunning || MidiBard.IsPlaying;
-
-        ImGuiUtil.PushIconButtonSize(new Vector2(ImGuiHelpers.GlobalScale * 40, ImGui.GetFrameHeight()));
-        if (!MidiBard.config.playOnMultipleDevices || (MidiBard.config.playOnMultipleDevices && MidiBard.config.usingFileSharingServices))
-        {
-            if (!ensembleRunning)
-            {
-                ImGui.BeginDisabled(isEnsembleButtonsDisabled);
-                if (ImGuiUtil.IconButton(FontAwesomeIcon.UserCheck, "btnEnsembleStart", ensemble_begin_ensemble_ready_check))
-                {
-                    if (MidiBard.config.UpdateInstrumentBeforeReadyCheck)
-                    {
-                        if (MidiBard.CurrentPlayback?.MidiFileConfig is { } config)
-                        {
-                            IPCHandles.UpdateMidiFileConfig(config);
-                        }
-
-                        if (!MidiBard.config.playOnMultipleDevices)
-                        {
-                            IPCHandles.UpdateInstrument(true);
-                        }
-                    }
-
-                    EnsembleManager.BeginEnsembleReadyCheck();
-                }
-                ImGui.EndDisabled();
-            }
-            else
-            {
-                if (ImGuiUtil.IconButton(FontAwesomeIcon.Stop, "btnEnsembleStop", ensemble_stop_ensemble))
-                {
-                    if (!MidiBard.config.playOnMultipleDevices)
-                    {
-                        IPCHandles.UpdateInstrument(false);
-                    }
-                    else
-                    {
-                        PartyChatCommand.SendClose();
-                    }
-                }
-            }
-
-            ImGui.SameLine();
-
-            ImGui.BeginDisabled(isEnsembleButtonsDisabled);
-            if (ImGuiUtil.IconButton(FontAwesomeIcon.Guitar, "UpdateInstrument", ensemble_update_instruments))
-            {
-                if (MidiBard.CurrentPlayback?.MidiFileConfig is { } config)
-                {
-                    IPCHandles.UpdateMidiFileConfig(config);
-                }
-
-                if (!MidiBard.config.playOnMultipleDevices)
-                {
-                    IPCHandles.UpdateInstrument(true);
-                }
-                else
-                {
-                    PartyChatCommand.SendUpdateInstrument();
-                }
-            }
-
-            if (ImGui.IsItemClicked(ImGuiMouseButton.Right))
-            {
-                if (!MidiBard.config.playOnMultipleDevices)
-                {
-                    IPCHandles.UpdateInstrument(false);
-                }
-                else
-                {
-                    PartyChatCommand.SendClose();
-                }
-            }
-            ImGui.EndDisabled();
-        }
-
-        //-------------------
-
-        ImGui.SameLine();
-        var muteButtonText = otherClientsMuted ? ensemble_unmute_other_clients : ensemble_mute_other_clients;
-        var muteButtonIcon = otherClientsMuted ? FontAwesomeIcon.VolumeMute : FontAwesomeIcon.VolumeUp;
-        if (ImGuiUtil.IconButton(muteButtonIcon, muteButtonText, muteButtonText))
-        {
-            // IsSndMaster => 0 = ON
-            // IsSndMaster => 1 = OFF
-            IPCHandles.SetOption("IsSndMaster", otherClientsMuted ? 0 : 1, false);
-            api.GameConfig.System.Set("IsSndMaster", 0);
-            otherClientsMuted ^= true;
-        }
-
-        //-------------------
-
-        ImGui.SameLine();
-        var muteLyricsButtonText = MidiBard.config.playLyrics ? "Disable lyrics" : "Enable lyrics";
-        var muteLyricsButtonIcon = MidiBard.config.playLyrics ? FontAwesomeIcon.Microphone : FontAwesomeIcon.MicrophoneSlash;
-        if (ImGuiUtil.IconButton(muteLyricsButtonIcon, muteLyricsButtonText, muteLyricsButtonText))
-        {
-            MidiBard.config.playLyrics = !MidiBard.config.playLyrics;
-            IPCHandles.SyncAllSettings();
-        }
-
-        //-------------------
-
-        ImGui.SameLine();
-        if (ImGuiUtil.IconButton(FontAwesomeIcon.WindowMinimize, "WindowMinimize", ensemble_minimize_other_clients))
-        {
-            IPCHandles.ShowWindow(Winapi.nCmdShow.SW_MINIMIZE);
-        }
-
-        //-------------------
-
-        if (!MidiFileConfigManager.UsingDefaultPerformer && !(MidiBard.config.playOnMultipleDevices && !MidiBard.config.usingFileSharingServices))
-        {
-            ImGui.SameLine();
-            if (ImGui.IsItemClicked(ImGuiMouseButton.Right))
-            {
-                IPCHandles.ShowWindow(Winapi.nCmdShow.SW_RESTORE);
-            }
-
-            //-------------------
-
-            ImGui.SameLine();
-            ImGui.BeginDisabled(isEnsembleButtonsDisabled);
-            if (ImGuiUtil.IconButton(FontAwesomeIcon.FolderOpen, "btnOpenConfigFolder", ensemble_open_midi_config_directory))
-            {
-                if (MidiBard.CurrentPlayback == null) return;
-
-                var fileInfo = MidiFileConfigManager.GetMidiConfigFileInfo(MidiBard.CurrentPlayback.FilePath);
-                var configDirectoryFullName = fileInfo.Directory.FullName;
-                // PluginLog.Debug(fileInfo.FullName);
-                // PluginLog.Debug(MidiBard.CurrentPlayback.FilePath);
-                // PluginLog.Debug(configDirectoryFullName);
-
-                Util.Extensions.OpenFolder(configDirectoryFullName);
-            }
-            ImGui.EndDisabled();
-
-            //-------------------
-
-            ImGui.SameLine();
-            ImGui.BeginDisabled(isEnsembleButtonsDisabled);
-            if (ImGuiUtil.IconButton(FontAwesomeIcon.Edit, "btnOpenConfigFile", ensemble_open_midi_config_file))
-            {
-
-                if (MidiBard.CurrentPlayback == null) return;
-
-                var fileInfo = MidiFileConfigManager.GetMidiConfigFileInfo(MidiBard.CurrentPlayback.FilePath);
-                // PluginLog.Debug(fileInfo.FullName);
-                // PluginLog.Debug(MidiBard.CurrentPlayback.FilePath);
-
-                Util.Extensions.OpenFile(fileInfo.FullName);
-            }
-            ImGui.EndDisabled();
-
-            //-------------------
-
-            ImGui.SameLine();
-            ImGui.BeginDisabled(isEnsembleButtonsDisabled);
-            if (ImGuiUtil.IconButton(FontAwesomeIcon.TrashAlt, "btnDeleteConfig", ensemble_delete_and_reset_current_file_config))
-            {
-                if (MidiBard.CurrentPlayback != null)
-                {
-                    MidiFileConfigManager.GetMidiConfigFileInfo(MidiBard.CurrentPlayback.FilePath).Delete();
-                    MidiBard.CurrentPlayback.MidiFileConfig = MidiFileConfigManager.GetMidiConfigFromTrack(MidiBard.CurrentPlayback.TrackInfos);
-                    MidiBard.CurrentPlayback.MidiFileConfig = BardPlayback.LoadDefaultPerformer(MidiBard.CurrentPlayback.MidiFileConfig);
-                    IPCHandles.UpdateInstrument(false);
-                }
-            }
-            ImGui.EndDisabled();
-
-            //-------------------
-
-            ImGui.SameLine();
-            ImGui.BeginDisabled(isEnsembleButtonsDisabled);
-            if (ImGui.Button(ensemble_save_default_performers))
-            {
-                MidiFileConfigManager.ExportToDefaultPerformer();
-            }
-            ImGui.EndDisabled();
-
-            //-------------------
-
-            // if (MidiBard.CurrentPlayback != null)
-            // {
-            //     ImGui.SameLine();
-
-            //     if (ImGuiUtil.IconButton(FontAwesomeIcon.WalkieTalkie, "##DoTest", "Report Loaded Playback"))
-            //     {
-            //         IPC.IPCHandles.ReportLoadedPlaybackInfo();
-            //     }
-            // }
-
-            ImGuiUtil.PopIconButtonSize();
-        }
-    }
-
-    private void DrawEnsembleControl()
-    {
-        if (!ShowEnsembleControlWindow) return;
-        if (!api.PartyList.IsPartyLeader()) return;
-
-        //ImGui.PushStyleVar(ImGuiStyleVar.FrameRounding, 2f);
-        //ImGui.PushStyleVar(ImGuiStyleVar.ItemSpacing, new Vector2(ImGui.GetStyle().ItemSpacing.X, ImGui.GetStyle().ItemSpacing.Y));
-        // ImGui.PushStyleColor(ImGuiCol.TitleBgActive, Theme.Components.WindowBg);
-        // ImGui.PushStyleColor(ImGuiCol.TitleBg, Theme.Components.WindowBg);
-        ImGui.PushStyleVar(ImGuiStyleVar.FramePadding, ImGui.GetStyle().FramePadding * 2.5f);
-        ImGui.PushStyleVar(ImGuiStyleVar.CellPadding, new Vector2(ImGui.GetStyle().CellPadding.Y));
-
-        if (ImGui.Begin(window_title_ensemble_panel + "###midibardEnsembleWindow", ref ShowEnsembleControlWindow))
-        {
-            DrawEnsembleMenu();
-
-            if (MidiFileConfigManager.UsingDefaultPerformer)
-            {
-                ImGui.SameLine();
-                ImGui.Text("[Using Default Performer]");
-            }
-
-            ImGui.Separator();
-            if (MidiBard.config.playOnMultipleDevices && !MidiBard.config.usingFileSharingServices)
-            {
-                ImGui.Button($"You are NOT using file sharing services to sync settings.\nTrack assign is disabled.\nPlease choose the tracks on clients individually.", new Vector2(-1, 100));
-            }
-            else if (MidiBard.CurrentPlayback == null)
-            {
-                if (ImGui.Button(ensemble_select_a_song_from_playlist, new Vector2(-1, ImGui.GetFrameHeight())))
-                {
-                    //try
-                    //{
-                    //    FilePlayback.LoadPlayback(new Random().Next(0, PlaylistManager.FilePathList.Count));
-                    //}
-                    //catch (Exception e)
-                    //{
-                    //    //
-                    //}
-                }
-            }
-            else
-            {
-                try
-                {
-                    var changed = false;
-                    var fileConfig = MidiBard.CurrentPlayback.MidiFileConfig;
-
-                    // use ensemble members config to define party selectbox order
-                    var partyList = api.PartyList.Select(partyMember => partyMember.GetPartyMemberData()).ToList();
-
-                    var cidToIndexMap = MidiBard.config.EnsembleMemberConfigs
-                        .Select((config, index) => new { config.Cid, Index = index })
-                        .ToDictionary(item => item.Cid, item => item.Index);
-
-                    var orderedPartyList = partyList
-                        .OrderBy(partyMember => cidToIndexMap.ContainsKey(partyMember.Cid)
-                                                ? cidToIndexMap[partyMember.Cid]
-                                                : int.MaxValue)
-                        .ToList();
-
-                    orderedPartyList.Insert(0, (Cid: 0, Name: "", World: ""));
-
-                    var partyNamesList = orderedPartyList
-                        .Select(partyMember => partyMember.Cid != 0 ? $"{partyMember.Name}·{partyMember.World}" : "")
-                        .ToArray();
-
-                    if (ImGui.BeginTable("fileConfig.Tracks", 4, ImGuiTableFlags.SizingFixedFit))
-                    {
-                        ImGui.TableSetupColumn("checkbox", ImGuiTableColumnFlags.WidthStretch, 1);
-                        ImGui.TableSetupColumn("instrument", ImGuiTableColumnFlags.WidthFixed);
-                        ImGui.TableSetupColumn("transpose", ImGuiTableColumnFlags.WidthFixed);
-                        ImGui.TableSetupColumn("playername", ImGuiTableColumnFlags.WidthStretch, 1.2f);
-
-                        var id = 125687;
-                        foreach (var dbTrack in fileConfig.Tracks)
-                        {
-                            ImGui.TableNextRow();
-                            ImGui.TableNextColumn();
-                            ImGui.PushID(id++);
-                            ImGui.PushStyleColor(ImGuiCol.Text, dbTrack.Enabled ? Theme.Components.Text : Theme.Components.TextDisabled);
-                            //var colUprLeft = dbTrack.Enabled ? Theme.Colors.Orange : Theme.Colors.Violet;
-                            //var pMin = GetWindowPos() + GetCursorPos();
-                            //var pMax = GetWindowPos() + GetCursorPos() + new Vector2(GetWindowContentRegionWidth(), GetFrameHeight());
-                            //GetWindowDrawList().AddRectFilledMultiColor(pMin, pMax, colUprLeft, 0, 0, colUprLeft);
-                            ImGui.AlignTextToFramePadding();
-                            changed |= ImGui.Checkbox($"{dbTrack.Index + 1:00} {dbTrack.Name}", ref dbTrack.Enabled);
-                            ImGui.TableNextColumn(); //1
-                            changed |= SelectInstrumentCombo($"##selectInstrument", ref dbTrack.Instrument);
-                            ImGui.TableNextColumn(); //2
-                            ImGui.SetNextItemWidth(ImGui.GetFrameHeight() * 3.3f);
-                            changed |= ImGuiUtil.InputIntWithReset($"##transpose", ref dbTrack.Transpose, 12, () => 0);
-                            ImGui.TableNextColumn(); //3
-                            ImGui.SetNextItemWidth(-1);
-
-                            var firstMidiFileCid = MidiFileConfig.GetFirstCidInParty(dbTrack);
-                            var selectedIdx = firstMidiFileCid == -1 ? 0 : orderedPartyList.FindIndex(i => i.Cid != 0 && i.Cid == firstMidiFileCid);
-
-                            if (ImGui.Combo("##partymemberSelect", ref selectedIdx, partyNamesList, partyNamesList.Length))
-                            {
-                                if (selectedIdx >= 1)
-                                {
-                                    var currentCid = orderedPartyList[selectedIdx].Cid;
-                                    if (firstMidiFileCid > 0 && currentCid != firstMidiFileCid)
-                                    {
-                                        // character changed, delete the old one
-                                        dbTrack.AssignedCids.Remove(firstMidiFileCid);
-                                        changed = true;
-                                    }
-
-                                    if (currentCid > 0)
-                                    {
-                                        // add character
-                                        if (!dbTrack.AssignedCids.Contains(currentCid))
-                                        {
-                                            dbTrack.AssignedCids.Insert(0, currentCid);
-                                            changed = true;
-                                        }
-                                    }
-                                }
-                                else
-                                {
-                                    // choose empty, remove all the characters in the same party
-                                    foreach (var member in api.PartyList)
-                                    {
-                                        if (dbTrack.AssignedCids.Contains(member.ContentId))
-                                        {
-                                            dbTrack.AssignedCids.Remove(member.ContentId);
-                                        }
-                                    }
-
-                                    changed = true;
-                                }
-                            }
-
-                            if (ImGui.IsItemClicked(ImGuiMouseButton.Right))
-                            {
-                                // choose empty, remove all the characters in the same party
-                                foreach (var member in api.PartyList)
-                                {
-                                    if (dbTrack.AssignedCids.Contains(member.ContentId))
-                                    {
-                                        dbTrack.AssignedCids.Remove(member.ContentId);
-                                    }
-                                }
-                                changed = true;
-                            }
-
-                            ImGuiUtil.ToolTip(ensemble_combo_tooltip_assign_track_character);
-
-                            ImGui.PopStyleColor();
-
-                            ImGui.PopID();
-                        }
-
-                        ImGui.EndTable();
-                    }
-
-                    if (changed)
-                    {
-                        fileConfig.Save(MidiBard.CurrentPlayback.FilePath);
-                        IPCHandles.UpdateMidiFileConfig(fileConfig);
-                    }
-                }
-                catch (Exception e)
-                {
-                    ImGui.TextUnformatted(e.ToString());
-                }
-            }
-
-            ImGui.Separator();
-            if (!MidiBard.config.playOnMultipleDevices)
-            {
-                ImGui.Checkbox(ensemble_config_update_instrument_when_begin_ensemble, ref MidiBard.config.UpdateInstrumentBeforeReadyCheck);
-
-            }
-#if DEBUG
-            try
-            {
-                foreach (var partyMember in api.PartyList)
-                {
-                    ImGui.TextUnformatted($"{partyMember.Name} {partyMember.ContentId:X} {partyMember.ObjectId:X} {partyMember.Address.ToInt64():X}");
-                    ImGui.SameLine();
-                    if (ImGui.SmallButton($"C##{partyMember.ContentId}"))
-                    {
-                        ImGui.SetClipboardText(partyMember.Address.ToInt64().ToString("X"));
-
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                ImGui.TextUnformatted(e.ToString());
-            }
-#endif
-        }
-
-        ImGui.End();
-
-        // ImGui.PopStyleColor(2);
-        ImGui.PopStyleVar(2);
-    }
-
-    private static bool SelectInstrumentCombo(string label, ref uint value)
-    {
-        var ret = false;
-        ImGui.PushStyleVar(ImGuiStyleVar.ItemSpacing, new Vector2(ImGui.GetStyle().FramePadding.Y));
-        if (value == 0)
-        {
-            ImGui.Image(TextureManager.Get(60042).GetWrapOrEmpty().ImGuiHandle, new Vector2(ImGui.GetFrameHeight()));
-        }
-        else
-        {
-            ImGui.Image(MidiBard.Instruments[value].IconTextureWrap.GetWrapOrEmpty().ImGuiHandle, new Vector2(ImGui.GetFrameHeight()));
-        }
-        if (ImGui.IsItemHovered()) ImGuiUtil.ToolTip(MidiBard.Instruments[value].InstrumentString);
-
-        ImGui.OpenPopupOnItemClick($"instrument{label}", ImGuiPopupFlags.MouseButtonLeft);
-
-        if (ImGui.BeginPopup($"instrument{label}"))
-        {
-            for (uint i = 1; i < MidiBard.Instruments.Length; i++)
-            {
-
-                ImGui.Image(MidiBard.Instruments[i].IconTextureWrap.GetWrapOrEmpty().ImGuiHandle, ImGuiHelpers.ScaledVector2(40, 40));
-                if (ImGui.IsItemClicked())
-                {
-                    value = i;
-                    ret = true;
-                    ImGui.CloseCurrentPopup();
-                }
-
-                if (ImGui.IsItemHovered())
-                {
-                    ImGuiUtil.ToolTip(MidiBard.Instruments[i].InstrumentString);
-                }
-
-                if (i is 4 or 9 or 14 or 19 or 23)
-                {
-                    continue;
-                }
-                ImGui.SameLine();
-            }
-            ImGui.EndPopup();
-        }
-
-        ImGui.PopStyleVar();
-
-        if (ImGui.IsItemHovered() && ImGui.IsMouseClicked(ImGuiMouseButton.Right))
-        {
-            value = 0;
-            ret = true;
-        }
-
-        return ret;
-    }
-}
