+++ Midibard/UI/TrackVisualizer/MidiPlot.cs
@@ -22,7 +22,6 @@ using System.Numerics;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 
-using Dalamud.Interface;
 using Dalamud.Interface.Colors;
 using Dalamud.Interface.Utility;
 
@@ -37,76 +36,13 @@ using MidiBard.Control;
 using MidiBard.Managers;
 using MidiBard.Util;
 
-using MidiBard2.Resources;
-
 using static Dalamud.api;
 
 namespace MidiBard;
 
 public partial class PluginUI
 {
-    private bool showTrackVisualizerWindow = false;
-    private bool _resetPlotWindowPosition = false;
-    private bool setNextLimit;
-    // private readonly double timeWindow = 10;
-    //private uint[] ChannelColorPalette = Enumerable.Range(0, 16).Select(i => ImGui.ColorConvertFloat4ToU32(HSVToRGB(i / 16f, 0.75f, 1))).ToArray();
-
-    public void ToggleTrackVisualizerWindow()
-    {
-        if (showTrackVisualizerWindow)
-            CloseTrackVisualizerWindow();
-        else
-            OpenTrackVisualizerWindow();
-    }
-
-    public void OpenTrackVisualizerWindow()
-    {
-        showTrackVisualizerWindow = true;
-    }
-
-    public void CloseTrackVisualizerWindow()
-    {
-        showTrackVisualizerWindow = false;
-    }
-
-    private void DrawPlotWindow()
-    {
-        if (!showTrackVisualizerWindow) return;
-
-        ImGui.PushStyleColor(ImGuiCol.TitleBg, Theme.Components.FrameBg);
-        ImGui.PushStyleColor(ImGuiCol.TitleBgActive, Theme.Components.FrameBg);
-
-        ImGui.PushStyleVar(ImGuiStyleVar.WindowPadding, -Vector2.One);
-        ImGui.SetNextWindowBgAlpha(0);
-        ImGui.SetNextWindowSize(ImGuiHelpers.ScaledVector2(640, 480), ImGuiCond.FirstUseEver);
-
-        if (_resetPlotWindowPosition)
-        {
-            ImGui.SetNextWindowPos(new Vector2(100), ImGuiCond.Always);
-            ImGui.SetNextWindowSize(ImGuiHelpers.ScaledVector2(640, 480), ImGuiCond.Always);
-            _resetPlotWindowPosition = false;
-        }
-
-        if (ImGui.Begin(Language.window_title_visualizor + "###midibardMidiPlot", ref showTrackVisualizerWindow, ImGuiWindowFlags.NoCollapse))
-        {
-            ImGui.PopStyleVar();
-            var icon = MidiBard.config.LockPlot ? FontAwesomeIcon.Lock : FontAwesomeIcon.LockOpen;
-            if (ImGuiUtil.AddHeaderIcon("lockPlot", icon.ToIconString(), Language.icon_button_tooltip_visualizer_follow_playback_tooltip))
-            {
-                MidiBard.config.LockPlot ^= true;
-            }
-            MidiPlotWindow();
-        }
-        else
-        {
-            ImGui.PopStyleVar();
-        }
-
-        ImGui.End();
-        ImGui.PopStyleColor(2);
-    }
-
-    private unsafe void MidiPlotWindow()
+    private unsafe void DrawMidiPlot()
     {
         if (ImGui.IsWindowAppearing())
         {
@@ -128,7 +64,7 @@ public partial class PluginUI
         }
         catch
         {
-            // silent fail
+            // ignored
         }
 
         string songName = "";
@@ -138,7 +74,7 @@ public partial class PluginUI
         }
         catch
         {
-            // silent fail
+            // ignored
         }
 
         //ImGui.SetCursorPos(ImGui.GetWindowContentRegionMin());
